made it easier to understand
